bot/requirements.txt
aiogram==2.25.1
tortoise-orm
requests
prodamuspy